Updated based on comments from Charles.
We are using GLM for RQ 4 instead of t-test as the depdendent variable is count (Poisson).

I created a new variable for weekend as I could not find how weekend/non-weekend has been defined in the code. I see wknd_ind but not sure how it is defined.

Everything else has been addressed.